Added `GetAliveTasks` RPC to coordination service.
This CL introduces a new `GetAliveTasks` RPC to the multi-controller JAX
coordination service. For a set of tasks `T`, `GetAliveTasks(T)` returns the
subset `A` of healthy tasks in `T`. To avoid hosts from disagreeing on which
tasks are healthy, `GetAliveTasks` has barrier-like semantics. In particular,
`GetAliveTasks` returns `A` only after every task in `A` has called
`GetAliveTasks(T)`. This API is intended to enable fault tolerant training
using multi-controller JAX.

Note that this CL introduces the `GetAliveTasks` API but it is not yet used. In
future CLs, I will pipe the API through to JAX and expose a `jax.alive_devices`
API. I'm separating the CLs to make the code easier to review.

# Implementation Details

`GetAliveTasks` has barrier-like semantics, and the coordination service
already implements barriers. However, the implementation of `GetAliveTasks`
differs from the barrier implementation in a couple of ways.

First, the barrier API expects a unique barrier name. For example, you might
call `barrier("foo")` to block on the barrier named "foo". For ergonomics,
`GetAliveTasks` does not require barrier names. Instead, the set of tasks
passed to `GetAliveTasks` acts like a name.

Second, the barrier API requires every barrier to have a unique name. It would
be annoying to construct a new name for every single barrier. Instead, the
barrier API allows programmers to repeatedly block on the same barrier.  To
implement this, every barrier has not only a name, but also a counter. For
example, you might request to block on the "foo.0" barrier or the "bar.4"
barrier. Coordination service clients keep track of the latest counter value to
use, incrementing it after every successful barrier. There is also some logic
surrounding how a failed client can recover the latest counter value.

Third, the barrier API allows barrier calls to time out. This means that a
barrier can enter a failed state, and all future entrants to the barrier should
observe this failed state. This introduces some complexity to the API, as
barrier state must be maintained even after the barrier has failed. The
`GetAliveTasks` does not implement timeouts and currently, a barrier cannot
fail. We can augment the API in the future if needed.

However, I would argue that the simpler semantics are sufficient, as there are
other sources of odd failure behavior in existing multi-controller JAX
programs. For example, if some tasks enter a collective and time out, while
other tasks enter the collective later, the program may hang or fail. Thus,
having stronger `GetAliveTasks` semantics increases the complexity of the API
without making multi-controller JAX programs meaningfully more fault tolerant.

PiperOrigin-RevId: 702840696